+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.7.0 - 2026-02-18
+### Updated
+- Make Wizard Home Page
+- Display versions of all contracts in the verification step.
+
+### Fixed
+- fix cache issue by adding option to clear it
+
 ## 0.6.0 - 2026-02-16
 ### Added
 - OKX support